Skip TestAccGkeonprem* (#8086)
* Skip TestAccGkeonprem*

* Add back destroy producers for tests
